tcdb-scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
cloudscraper>=1.2.71
browser-cookie3>=0.19.1
pytest>=7.0.0
